Update NewtonManager.cpp
the steptime was set to 1/1000 : aMaxphysicfps(1.0f / 1000.0f) which is adapted to tune the control laws.
The solver was set to 0 :   NewtonSetSolverIterations(nWorld, 0); which is the correct mode for the exact solver.
The time ellapsed was output in the file 'history.txt' to check whether the physictimestep was close to the real time spent between 2 successive updates.