muscle implementation
OXVehicleRayCast.h is the Listener for the ragdoll manager. All the bodies and joints of the humanoid are designed as a tree structure and have unique Callbacks triggered by this manager.
the Muscle class is implemented in biped.h
The LineDebugManager class deals with GLlines creation and updates. It is implemented in WindowGL.h and it is called in the Muscle class to set up the graphic representation of the muscle in the form of lines.