Optimization parameter loaded from XML
The parameters to be optimized are loaded from OptimizationParameters.xml using TinyXML.

Add LoadXML.h in your solution headers.
When loaded, the parameters are scaled to be sent to the optimization algorithm (TO DO). Then the scaling is removed to run NewtoDynamics simulation.

next: include CMAS routine in this solution.